Compatibility with Node v6.x.
In Node.js v6.x, hash.update assumes the input to be utf8. Previous versions of Node.js assumes the input to be in binary. This causes hashes generated to be incompatible between Node.js < 6.x and >= 6.x. Make the call to hash.update explicit so that hashes can continue to be compatible.